refactor: move gene_incomplete_random_walk_dataset and gene_physionet2012 to data.generating, delete tests/unified_data_for_test;
